requirements.txt
streamlit>=1.31
openai>=1.12
pandas>=2.2
python-dotenv>=1.0 